Move treelist and treedict to pymc3.util
